Different setup for django python shell
Originally when calling the Django Python Shell, it would setup a shell
inside of the shell. If done this way, then code completion from some
packages such as `company-capf` would not work. This is a fixed for that
issue. Environment can be setup on the same shell instead of running one
inside the other \o/